Update minimal-mistakes and document installing with and without docker.
This commit makes the following changes:
* Adds some documentation - Install.md
* Changes the way the local copy is build (both with and without Docker)
* Removes the "in-repo" copy of the minimal mistakes jekyll theme

Documentation
-------------

I wrote a hasty install guide - Install.md. I suggest at a later date,
we merge this into website2016.wiki, but for now I leave it here to not
detatch it from this commit.

It would be great if someone could verify this method, and we still need
to think about documentation for other platforms.

Gemfile changes
---------------

Currently, `jekyll` is provided through the Docker container (I think);
this effectively makes docker a requirement. This commit changes the
Gemfile to grab a local copy of jekyll.

* Non-docker will now download a copy of jekyll from rubygems and use
it to build the website
* Docker builds will still succeed, but the jekyll from the container
will effectively be superceded by the rubygems version.

Remove unedited minimal-mistakes to use the gem version
-------------------------------------------------------

The minimal mistakes jekyll theme is now provided as a gem. This commit
removes all the theme content, excepting those files that have been
edited: edited files will supercede the gem's version.

The justification for this was I wanted minimal-mistakes to be easier to
update. I also wanted everything to be runnable locally, without Docker.

I am unlikely to PR this branch.

Discussion points:
* Should we use minimal-mistakes as a gem, or periodically update it by
pulling from its repo?
* Could a docker-lover chip in some thoughts about whether this
re-downloading of jekyll (once as a container; once installing the gem)
is a terrible sin.